Update dependencies, refactor components, and enhance styles
- Added autoprefixer and postcss to package.json and pnpm-lock.yaml.
- Refactored postcss.config.mjs to use object syntax for plugins.
- Updated tsconfig.json to include tailwind.config.ts.
- Introduced ShimmerCard and ShimmerGrid components for loading states.
- Replaced LoadingCard with ShimmerCard in tests and components.
- Adjusted styles in CryptoCardContent and hover effects for better UI consistency.
- Reduced sleep duration in asset fetching logic for improved performance.
