+++ components/crypto/__tests__/LoadingCard.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react'
 import '@testing-library/jest-dom'
-import LoadingCard from '../../ui/loading-card'
+import ShimmerCard from '../../ui/shimmer-card'
 
 // Mock LoadingSparkline
 jest.mock('../loading-sparkline', () => ({
@@ -14,39 +14,31 @@ jest.mock('../loading-sparkline', () => ({
 
 describe('LoadingCard', () => {
   it('should render with proper structure', () => {
-    render(<LoadingCard />)
+    render(<ShimmerCard />)
 
     // Check for Card container
     const card = screen.getByText('Loading Sparkline').closest('[data-slot="card"]')
     expect(card).toBeInTheDocument()
     expect(card).toHaveClass('glassmorphic')
     expect(card).toHaveClass('animate-pulse')
-    expect(card).toHaveClass('min-h-[210px]')
   })
 
   it('should have glassmorphic styling', () => {
-    const { container } = render(<LoadingCard />)
+    const { container } = render(<ShimmerCard />)
     
     const card = container.querySelector('.glassmorphic')
     expect(card).toBeInTheDocument()
   })
 
   it('should have animate-pulse class', () => {
-    const { container } = render(<LoadingCard />)
+    const { container } = render(<ShimmerCard />)
     
     const card = container.querySelector('.animate-pulse')
     expect(card).toBeInTheDocument()
   })
 
-  it('should have proper minimum height', () => {
-    const { container } = render(<LoadingCard />)
-    
-    const card = container.querySelector('.min-h-\\[210px\\]')
-    expect(card).toBeInTheDocument()
-  })
-
   it('should render skeleton elements', () => {
-    const { container } = render(<LoadingCard />)
+    const { container } = render(<ShimmerCard />)
     
     // Check for skeleton elements with opacity
     const skeletonElements = container.querySelectorAll('.opacity-30')
@@ -54,7 +46,7 @@ describe('LoadingCard', () => {
   })
 
   it('should render LoadingSparkline with correct opacity', () => {
-    render(<LoadingCard />)
+    render(<ShimmerCard />)
     
     const loadingSparkline = screen.getByTestId('loading-sparkline')
     expect(loadingSparkline).toBeInTheDocument()
@@ -62,14 +54,14 @@ describe('LoadingCard', () => {
   })
 
   it('should have flex layout structure', () => {
-    const { container } = render(<LoadingCard />)
+    const { container } = render(<ShimmerCard />)
     
     const card = container.querySelector('.flex.flex-col.justify-between')
     expect(card).toBeInTheDocument()
   })
 
   it('should render header section with skeleton elements', () => {
-    const { container } = render(<LoadingCard />)
+    const { container } = render(<ShimmerCard />)
     
     // Check for header section
     const headerSection = container.querySelector('.flex.items-center.justify-between')
@@ -81,7 +73,7 @@ describe('LoadingCard', () => {
   })
 
   it('should render icon placeholder', () => {
-    const { container } = render(<LoadingCard />)
+    const { container } = render(<ShimmerCard />)
     
     // Check for circular icon placeholder
     const iconPlaceholder = container.querySelector('.w-8.h-8.bg-gray-400.rounded-full')
@@ -89,7 +81,7 @@ describe('LoadingCard', () => {
   })
 
   it('should render price section with skeleton', () => {
-    const { container } = render(<LoadingCard />)
+    const { container } = render(<ShimmerCard />)
     
     // Check for price skeleton in the bottom section
     const priceSkeleton = container.querySelector('.h-6.bg-gray-400.rounded')
