+++ lib/data/assets.ts
@@ -30,7 +30,7 @@ export async function fetchAssets(): Promise<Array<Asset>> {
     const { data: assets } = await res.json();
     console.log("Fetched assets:", assets?.length || 0);
 
-    await sleep(3000);
+    await sleep(1000);
     
     return assets.map((asset: Asset) => ({
       ...asset,
