+++ package.json
@@ -35,11 +35,13 @@
     "@types/node": "^20",
     "@types/react": "^19",
     "@types/react-dom": "^19",
+    "autoprefixer": "^10.4.21",
     "eslint": "^9",
     "eslint-config-next": "15.3.2",
     "jest": "^30.0.4",
     "jest-environment-jsdom": "^30.0.4",
-    "tailwindcss": "^4",
+    "postcss": "^8.5.6",
+    "tailwindcss": "^4.1.11",
     "ts-jest": "^29.4.0",
     "tw-animate-css": "^1.3.0",
     "typescript": "^5"
